RIP-319, dedicated /egrades controller; convert_per_grading_basis util w/ pnp_cutoff
