+++ ripley/lib/canvas_utils.py
@@ -23,13 +23,15 @@ SOFTWARE AND ACCOMPANYING DOCUMENTATION, IF ANY, PROVIDED HEREUNDER IS PROVIDED
 ENHANCEMENTS, OR MODIFICATIONS.
 """
 
+from itertools import groupby
 import re
 
 from flask import current_app as app
 from ripley.api.errors import InternalServerError, ResourceNotFoundError
 from ripley.externals import canvas, data_loch
 from ripley.externals.s3 import upload_dated_csv
-from ripley.lib.berkeley_course import course_section_name
+from ripley.lib.berkeley_course import course_section_name, course_to_api_json, section_to_api_json, \
+    sort_course_sections
 from ripley.lib.berkeley_term import BerkeleyTerm
 from ripley.lib.sis_import_csv import SisImportCsv
 from ripley.lib.util import utc_now
@@ -104,6 +106,77 @@ def format_term_enrollments_export(term_id):
     return f"{term_id.replace(':', '-')}-term-enrollments-export"
 
 
+def get_official_sections(canvas_site_id):
+    canvas_sections = canvas.get_course_sections(canvas_site_id)
+    canvas_sections = [canvas_section_to_api_json(cs) for cs in canvas_sections if cs.sis_section_id]
+    canvas_sections_by_id = {cs['id']: cs for cs in canvas_sections if cs['id']}
+    section_ids = list(canvas_sections_by_id.keys())
+    term_id = canvas_sections[0]['termId']
+    sis_sections = sort_course_sections(
+        data_loch.get_sections(term_id, section_ids) or [],
+    )
+    if len(sis_sections) != len(section_ids):
+        app.logger.warn(f'Canvas site ID {canvas_site_id} has {len(section_ids)} sections, but SIS has {len(sis_sections)} sections.')
+
+    def _section(section_id, rows):
+        canvas_section = canvas_sections_by_id[section_id]
+        return {
+            **canvas_section,
+            **section_to_api_json(rows[0], rows[1:]),
+        }
+    official_sections = []
+    for section_id, rows in groupby(sis_sections, lambda s: s['section_id']):
+        official_sections.append(_section(section_id, list(rows)))
+    return official_sections, section_ids, sis_sections
+
+
+def get_teaching_terms(current_user, section_ids, sections):
+    berkeley_terms = BerkeleyTerm.get_current_terms()
+    canvas_terms = [term.sis_term_id for term in canvas.get_terms() if term.sis_term_id]
+    terms = []
+    for key, term in berkeley_terms.items():
+        if term.to_canvas_sis_term_id() not in canvas_terms:
+            continue
+        if key != 'future' or term.season == 'D':
+            terms.append(term)
+
+    teaching_sections = []
+    if (current_user.is_teaching or current_user.canvas_masquerading_user_id):
+        instructor_uid = current_user.uid
+        teaching_sections = sort_course_sections(
+            data_loch.get_instructing_sections(instructor_uid, [t.to_sis_term_id() for t in terms]) or [],
+        )
+    if not len(teaching_sections):
+        teaching_sections = sections
+    courses_by_term = {}
+    for section_id, sections in groupby(teaching_sections, lambda s: s['section_id']):
+        sections = list(sections)
+        section = next((s for s in sections if s.get('is_co_instructor', False) is False), None)
+        co_instructor_sections = [s for s in sections if s.get('is_co_instructor', True) is True]
+        course_id = section['course_id']
+        term_id = section['term_id']
+        if term_id not in courses_by_term:
+            courses_by_term[term_id] = {}
+        if course_id not in courses_by_term[term_id]:
+            term = BerkeleyTerm.from_sis_term_id(term_id)
+            courses_by_term[term_id][course_id] = course_to_api_json(term, section)
+        courses_by_term[term_id][course_id]['sections'].append({
+            **section_to_api_json(section, co_instructor_sections),
+            'isCourseSection': section_id in section_ids,
+        })
+
+    def _term_courses(term_id, courses_by_id):
+        term = BerkeleyTerm.from_sis_term_id(term_id)
+        return {
+            'classes': list(courses_by_id.values()),
+            'name': term.to_english(),
+            'slug': term.to_slug(),
+            'termId': term_id,
+            'termYear': term.year,
+        }
+    return [_term_courses(term_id, courses_by_id) for term_id, courses_by_id in courses_by_term.items()]
+
+
 def prepare_egrade_export(course):
     app.logger.warning(f'E-Grades job started for course {course.id}')
     official_grades = []
@@ -117,18 +190,6 @@ def prepare_egrade_export(course):
             'uid': user.login_id if hasattr(user, 'login_id') else None,
         })
     return official_grades
-
-
-def _extract_grades(enrollments):
-    keys = ['current_score', 'current_grade', 'final_score', 'final_grade override_score', 'override_grade']
-    grade_hash = dict.fromkeys(keys, None)
-    enrollments = filter(lambda e: e.type == 'StudentEnrollment' and hasattr(e, 'grades'), enrollments)
-    if enrollments:
-        api_grades = enrollments[0].grades
-        for key in keys:
-            if key in api_grades:
-                grade_hash[key] = api_grades[key]
-    return grade_hash
 
 
 def sis_enrollment_status_to_canvas_course_role(sis_enrollment_status):
@@ -214,6 +275,18 @@ def _canvas_site_term_json(canvas_site):
                 'name': term.to_english(),
             }
     return api_json
+
+
+def _extract_grades(enrollments):
+    keys = ['current_score', 'current_grade', 'final_score', 'final_grade override_score', 'override_grade']
+    grade_hash = dict.fromkeys(keys, None)
+    enrollments = filter(lambda e: e.type == 'StudentEnrollment' and hasattr(e, 'grades'), enrollments)
+    if enrollments:
+        api_grades = enrollments[0].grades
+        for key in keys:
+            if key in api_grades:
+                grade_hash[key] = api_grades[key]
+    return grade_hash
 
 
 def _update_enrollments_in_background(sis_term_id, course, all_sections, deleted_section_ids, sis_import):
