+++ tests/test_api/test_canvas_site_controller.py
@@ -304,71 +304,6 @@ class TestCanvasSiteProvisionSections:
             assert course['slug'] == 'astron-218-B-2023'
 
 
-class TestEgradeExportOptions:
-
-    @classmethod
-    def _api_egrade_export_options(cls, client, expected_status_code=200, failed_assertion_message=None):
-        response = client.get('/api/canvas_site/egrade_export/options')
-        assert response.status_code == expected_status_code, failed_assertion_message
-        return response.json
-
-    def test_unauthorized(self, client, fake_auth):
-        """Denies unauthorized users."""
-        users = {
-            'anonymous': None,
-            'no_canvas_account': no_canvas_account_uid,
-            'not_enrolled': not_enrolled_uid,
-            'reader': reader_uid,
-            'student': student_uid,
-            'teaching_assistant': ta_uid,
-        }
-        for user_type, uid in users.items():
-            if uid:
-                fake_auth.login(canvas_site_id='8876542', uid=no_canvas_account_uid)
-            self._api_egrade_export_options(
-                client,
-                expected_status_code=401,
-                failed_assertion_message=f'Unexpected response status for {user_type} user',
-            )
-
-    def test_teacher(self, client, app, fake_auth):
-        """Allows teacher."""
-        with requests_mock.Mocker() as m:
-            canvas_site_id = '8876542'
-            register_canvas_uris(app, {
-                'account': ['get_admins', 'get_terms'],
-                'course': [
-                    f'get_by_id_{canvas_site_id}',
-                    f'get_sections_{canvas_site_id}',
-                    f'get_settings_{canvas_site_id}',
-                    'get_enrollments_4567890',
-                ],
-                'user': [f'profile_{teacher_uid}'],
-            }, m)
-            fake_auth.login(canvas_site_id=canvas_site_id, uid=teacher_uid)
-            api_json = self._api_egrade_export_options(client)
-            # Verify
-            assert api_json['gradingStandardEnabled'] is True
-
-    def test_admin(self, client, app, fake_auth):
-        """Allows teacher."""
-        with requests_mock.Mocker() as m:
-            canvas_site_id = '8876542'
-            register_canvas_uris(app, {
-                'account': ['get_admins', 'get_terms'],
-                'course': [
-                    f'get_by_id_{canvas_site_id}',
-                    f'get_sections_{canvas_site_id}',
-                    f'get_settings_{canvas_site_id}',
-                ],
-                'user': [f'profile_{admin_uid}'],
-            }, m)
-            fake_auth.login(canvas_site_id=canvas_site_id, uid=admin_uid)
-            api_json = self._api_egrade_export_options(client)
-            # Verify
-            assert api_json['gradingStandardEnabled'] is True
-
-
 class TestGetRoster:
 
     def test_anonymous(self, client):
